ML Deployment in AWS
This project depicts how we can train, build and deploy in AWS SageMaker. 
We had taken the Marketing Banking data from the UCI repository and stored it in S3 buckets. Then this was used to train using the XGBoost algorithm. Finally we got the endpoint for the model in AWS SageMaker.