Maximize the Profit Margin for Diamonds
Objective
You own a diamond business and you have to replenish your inventory. You have $5,000,000 budgeted for the purchase of new diamonds. You have a historical list of diamonds that were purchased from various distributor vendors and the prices they retailed for. You also have a list of diamonds that are currently on the market, but this list does not include any set prices. Instead, you will place an offer for a diamond and the distributor will make a single yes-or-no decision to sell you the diamond at that price. If your offer is accepted then you receive the diamond and you will be awarded the difference between the retail price your company can eventually sell the diamond for and the price that you offered as profit. If you fail to make a purchase, you simply do not get anything for profit.
The offer column should be populated only on diamonds you wish to make an offer on, and the sum of the offers needs to be less than or equal to $5,000,000.

Directions
You have been given two data sets. The dataset labeled “training.csv” contains historical data regarding how much diamonds have sold for; both the retail price the diamond sells at and the price paid to the vendor to acquire it. You can use this data to train any models you see fit to answer the question. The second data set labeled “offers.csv” is the dataset you will provide offers for. Note neither the retail price or the price you would need to pay is present in the offers file.

The given solution contains the python code file which calculates the offer & retail price on the test data set. The actual diamonds on which the offer price were made were selected using Excel. Tried working around the Linear Programming code in Python but somehow it didn't work.